SPARK-34369: Refactored the Join Metrics to a different test Suite as per the review comments.
